Version 1.2
If you use ConvLib, an NPC will be automatically paused, but this has to be after ConvLib!
You can add an "npc_using" script to an exit to have that run instead of the standard phrase for moving. It has access to a local variable "npc".
If game.npcdebug is set to true, you will see debugging info in blue; this will tell you what NPCs are dong and planning to do, which should help you work out why they are not behaving.
There are also groups now; NPCs can join groups and act together whilst in the group.